// rotaplanner/generated/client/types.gen.ts
export type Activity = {
    /**
     * Id
     */
    id: string;
    /**
     * Name
     */
    name: string;
    /**
     * Location
     */
    location: string | null;
    /**
     * Timeslots
     */
    timeslots?: Array<Timeslot>;
    /**
     * Activity Start
     */
    readonly activity_start: string;
    /**
     * Activity Finish
     */
    readonly activity_finish: string;
};

/**
 * DateRange
 */
export type DateRange = {
    /**
     * Start
     */
    start: string;
    /**
     * End
     */
    end: string;
};

/**
 * Location
 */
export type Location = {
    /**
     * Id
     */
    id: string | null;
    /**
     * Name
     */
    name: string;
};

/**
 * Staff
 */
export type Staff = {
    /**
     * Id
     */
    id: string | null;
    /**
     * Name
     */
    name: string;
};
// ...
export type AvailableForTimeslotResult = {
    /**
     * Availability Type
     */
    availability_type: string;
    /**
     * Available Staff
     */
    available_staff: Array<Staff>;
};
// ...
export type GetTableDataResponses = {
    /**
     * TableDataResult
     */
    200: {
        /**
         * Queryversion
         */
        queryVersion?: string;
        dateRange: DateRange;
        /**
         * Staff
         */
        staff: Array<string>;
        /**
         * Locations
         */
        locations: Array<string>;
        /**
         * Staffdata
         */
        staffData: {
            [key: string]: Staff;
        };
        /**
         * Locationsdata
         */
        locationsData: {
            [key: string]: Location;
        };
        /**
         * Activities
         */
        activities: Array<Activity>;
    };
};

export type GetTableDataResponse = GetTableDataResponses[keyof GetTableDataResponses];